requirements.txt
openai
mysql-connector-python
pandas
numpy
yfinance
requests
pandas-ta 